Create Kill.java
Description : 

A war is happening. All the soldiers are sitting in a circle. The soldier at the first position has the sword. He killed the second soldier and passed the sword to the third solider. The third soldier killed the fourth soldier and passed the sword to the fifth solider and so on.

At the end only one person will remain alive in the circle. What is the position of the person who is alive.

Take the no of soldiers from the user input. The soldiers are represented with i where 1<=i<=n+1 where n is the size of the array.